Build flex when build starts
The flex binary will be compiled at the very beginning
so that when it is needed it's available for use.

The reason for compiling it from source rather than
using a precompiled version is simply due to the fact
that the old flex binary shipped in AOSP does not play
well with newer systems, as well as newer binaries
not being compatible with AOSP source code.

To mitigate this issue, we can compile flex 2.5.39
from source which will work fine that way.

Change-Id: I8efa4348aab929aa0551bc86e93db9057bc8eec2