Add support for overriding security patch
In the case that we pick the patches from the bulletin, we only change
the patch level to *-01 and not *-05 since the *-05 bulletins only
contain vendor-specific patches that are either proprietary or in the
kernel. That is because we don't have the ability to control these patches
since maintainers need to rely on the oem, odm and SoC vendor to update
these components.

For maintainers that are able to get free (libre) components up to date,
this patch adds support for overriding the security patch level.

In order to override the security patch level, define the variable
PLATFORM_SECURITY_PATCH_OVERRIDE in the aosp_*.mk of your device
tree.

For example:

PLATFORM_SECURITY_PATCH_OVERRIDE := 2024-07-05

At the moment, no date comparison is performed, so make sure you update
this whenever the original platform security patch changes. A change may
be added in the future that checks for this and uses the level that is
more up to date.